Former-commit-id: 9f282be3f3fa5ee2ce9e0f6c2b9d63f807c18f3f

+++ libraries/reconstruction/ml2d.h
@@ -79,7 +79,7 @@ class ModelML2D
 {
 public:
     /** Number of reference images */
-    size_t n_ref;
+    int n_ref;
     /** References images */
     std::vector < Image<double> > Iref, WsumMref;
     /** Sigma value for expected pixel noise */
